+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Bugfix) (Platform) Fix topology for Gateways
 - (Feature) (Platform) Dump CLI switch to Services
 - (Feature) (Platform) Fix ImagePullSecrets Merge
+- (Feature) (Platform) Update Failed Releases
 
 ## [1.3.2](https://github.com/arangodb/kube-arangodb/tree/1.3.2) (2025-11-20)
 - (Bugfix) (Platform) Increase memory limit for Inventory